requirement.txt
mysqlclient
django
djangorestframework
djangorestframework-simplejwt
django-filter
pillow